% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_bin_hourly.R
\name{binHourly}
\alias{binHourly}
\title{Bin hourly}
\usage{
binHourly(df, long)
}
\arguments{
\item{df}{dataframe to be binned hourly}

\item{long}{if long == T, data in long format. If long==F, data in wide format}
}
\value{
binned dataframe, I guess
}
\description{
bins hourly by median. APparently. I have no idea where this is used currently,
might be an artifact
}
\examples{
none
}
