% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_calculate_GR_classic.R
\name{calculate_GR_classic}
\alias{calculate_GR_classic}
\title{Calculate GR classic}
\usage{
calculate_GR_classic(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
A list with the first element containing growthrates in "wide" format, 
the second elmenent containing growthrates in "long" format
}
\description{
Calculates growthrates the "classical" way, by binning a sizerange and applying linear
fits to each bin.
}
\examples{
none
}
