% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_combine_acsm_hom.R
\name{combine_acsm_hom}
\alias{combine_acsm_hom}
\title{Combine ACSM and HOM data}
\usage{
combine_acsm_hom(acsmdata, homdata, averagetime)
}
\arguments{
\item{acsmdata}{ACSM data}

\item{homdata}{HOM data}
}
\value{
a combined data frame
}
\description{
Combines ACSM data with HOM data after first getting the hourly median of each.
}
\examples{
none
}
