% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_combine_acsm_SA_HOM_CS_kappa.R
\name{combine_acsm_SA_HOM_CS_kappa}
\alias{combine_acsm_SA_HOM_CS_kappa}
\title{combine ACSM, SA, HOM and CS kappa data}
\usage{
combine_acsm_SA_HOM_CS_kappa(acsm_data_long, HOM_data, SA_data,
  CS_kappa_data)
}
\arguments{
\item{acsm_data_long}{acsm data, long format}

\item{HOM_data}{HOM data, long format}

\item{SA_data}{SA data, long format}

\item{CS_kappa_data}{CS kappa data, long format}
}
\value{
combined data frame (long format)
}
\description{
Combines all into long format
}
\examples{
combine_acsm_SA_HOM_CS_kappa(acsm_data_long,HOM_data,SA_data,CS_kappa_data)
}
