% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_convertToLimit.R
\name{convertToLimit}
\alias{convertToLimit}
\title{Convert to limit}
\usage{
convertToLimit(x, i)
}
\arguments{
\item{x}{the character/factor value/vector from which limits get extracted}

\item{i}{flag. If i = 0, the lower limit gets extracted. If i = 1, the higher limit.}
}
\value{
the desired limit
}
\description{
utility function that obtains factor limits from its respective character vector,
obtained by the cut function.
}
\details{
For example: (0,5] could indicate the range "x >0, x <= 5". convertToLimit("(0,5]",0)
would then return 0 and convertToLimit("(0,5]",1) would return 5
}
\examples{
convertToLimit("(0,5]",0)
convertToLimit("(0,5]",5)
}
