% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_correctDP.R
\name{correctDP}
\alias{correctDP}
\title{Correct particle diamater}
\usage{
correctDP(dp)
}
\arguments{
\item{dp}{particle diameter in nanometers}
}
\value{
corrected particle diameter
}
\description{
Uses the correction function found by Joonas Enroth for correction NAIS particle diamater
}
\details{
Correction factor is: dp_corr=(10^(log10(dp)*(1/1.073))) . Have to double check this
}
\examples{
correctDP(2.42)
}
