% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_create_splined_trajectories_V2.R
\name{create_splined_trajectories_V2}
\alias{create_splined_trajectories_V2}
\title{create splined trajectories}
\usage{
create_splined_trajectories_V2(modesizeframe, eventIDframe, sparfac)
}
\arguments{
\item{modesizeframe}{dataframe containing timeseries information of the growing mode}

\item{eventIDframe}{dataframe containing information about the eventIDs (used for selecting "good" events)}

\item{sparfac}{#the "stiffness" of the trajectories /smoothing factor. Between 0 and 1.
0 creates "wiggly trajectories", 1 very stiff one. Values around 0.95 seem to be good}
}
\value{
a list: 1 element: modesizeframe with some additional columns, for example the
smoothing spline evaluated in the initial timesteps. \Cr
2 element: Another list containing as first element the eventID's of the growing event and as
second element the smoothing spline functions for the eventID's in the first element.
}
\description{
create splined trajectories
}
\examples{
NONE
}
