% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_evaluate_splined_trajectories.R
\name{evaluate_splined_trajectories}
\alias{evaluate_splined_trajectories}
\title{Evaluate splined trayectories}
\usage{
evaluate_splined_trajectories(splined.smooth_functions, filt.tray.df,
  timestep)
}
\arguments{
\item{splined.smooth_functions}{a file containing the functions}

\item{filt.tray.df}{a file containing the filtered trajectories}

\item{timestep}{timestep of function evaluation in seconds.}
}
\value{
a dataframe containing the evaluation results
}
\description{
Evaluate splined trayectories
}
\details{
filt.tray.df is only used to get the starting times for the trajectories
}
\examples{
NONE
}
