% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_get_MAD_distance.R
\name{get_MAD_distance}
\alias{get_MAD_distance}
\title{Calculate MAD distance}
\usage{
get_MAD_distance(vec, how)
}
\arguments{
\item{vec}{The base vector for the MAD distance calculation. For each element in the vector
MAD distance respective to this vector will be calculated}

\item{how}{"single" or "double"}
}
\value{
a vector containing the MAD distances for elements in the vector
}
\description{
Calculates the MAD (median absolute derivation) distance for each point in a vector. \cr
This helps in detecting outliers in non-standard distributions
}
\details{
It can be single sided (treat all values equally) or
double sided (treat values < median separately from values > median) \cr
I recommend to use double sided.
}
\examples{
get_MAD_distance(c(5,120,42,3,1,1,1,4,-3,6),"double")
}
