% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getCharge_eventID.R
\name{getCharge_eventID}
\alias{getCharge_eventID}
\title{Get charge from event ID}
\usage{
getCharge_eventID(eventID)
}
\arguments{
\item{eventID}{An eventID of the form "Pos123","Neg42",etc}
}
\value{
The charge of the event. "positive" or "negative"
}
\description{
EventIDs are given to positive and negative ion growing events. For example: Pos101 is the
growing event 101 of a positive ion. This function extracts the charge, "positive" in this case.
}
\examples{
getCharge_eventID("Pos101")
}
