% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getInd_eventID.R
\name{getInd_eventID}
\alias{getInd_eventID}
\title{Get event ID index}
\usage{
getInd_eventID(eventID)
}
\arguments{
\item{eventID}{An event ID of form "Pos42","Neg203",etc}
}
\value{
The index as a character
}
\description{
Gets the index of an event ID of the form "Pos42". In this case, it would return "42"
}
\examples{
getInd_eventID("Neg12")
}
