% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ggplot_Regression.R
\name{ggplotRegression}
\alias{ggplotRegression}
\title{ggplot Regression}
\usage{
ggplotRegression(fit)
}
\arguments{
\item{fit}{:A fit created with lm()}
}
\value{
A ggplot object
}
\description{
plot a linear regression fit with ggplot
}
\examples{
fit=lm(x,y,mydata)
ggplotRegression(fit)

}
