% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_acsm_dat.R
\name{import_acsm_dat}
\alias{import_acsm_dat}
\title{import_Acsm_dat}
\usage{
import_acsm_dat(where, filename)
}
\arguments{
\item{where}{path to the file}

\item{filename}{name of the file with extension}
}
\value{
data table containing the data
}
\description{
Imports acsm data for the bolivia campaign
}
\examples{
importAcsmData()
}
