% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_CS_kappa_dat.R
\name{import_CS_kappa_dat}
\alias{import_CS_kappa_dat}
\title{Imports CS kappa data}
\usage{
import_CS_kappa_dat(where, filename)
}
\value{
a data table with the data
}
\description{
imports kappa corrected condensation sink data, created by Joonas Enroth
}
\details{
Issues: Timezone not sure. Have to verify
}
\examples{
import_CS_kappa_data
}
