% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Import_GR_dat.R
\name{import_GR_mode_dat}
\alias{import_GR_mode_dat}
\title{import GR mode data}
\usage{
import_GR_mode_dat(where, writeFile = FALSE)
}
\arguments{
\item{where}{path to folder containing the files}

\item{writeFile}{if TRUE, write to file.}
}
\value{
data frame with the growing mode information
}
\description{
imports data for the size growing of the growing mode in time. Saltena campaign.
Can save to file if desired.
All files within the folder will be loaded
}
\examples{
to be filled
}
