% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_HOM_dat.R
\name{import_HOM_dat}
\alias{import_HOM_dat}
\title{Import HOM data}
\usage{
import_HOM_dat(where, filename)
}
\arguments{
\item{where}{path to the file}

\item{filename}{filename with extension}
}
\value{
data in data table
}
\description{
imports HOM data created with apitof
}
\examples{
import_HOM_dat(where,"homdata.csv")
}
