% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_meteo_dat.R
\name{import_meteo_dat}
\alias{import_meteo_dat}
\title{Import meteo data}
\usage{
import_meteo_dat(where, filename)
}
\arguments{
\item{where}{path to the folder conaining the file}

\item{filename}{filename with extension}
}
\value{
returns data in data table
}
\description{
Imports meteorological data for the saltena campaign
}
\examples{
import_meteo_dat()
}
