% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_NAIS_data_2013_2018.R
\name{import_NAIS_data_2013_2018}
\alias{import_NAIS_data_2013_2018}
\title{Imports NAIS}
\usage{
import_NAIS_data_2013_2018(where, year)
}
\arguments{
\item{where}{Path to the containing folder. All files with extension .sum within the path will be pulled.}

\item{year}{The year to be imported. (You can import one year only)}
}
\value{
data table of data
}
\description{
Import NAIS data (.sum) from the NAIS3 in CHC
}
\details{
These files are big, it might take a while.
}
\examples{
import_NAIS_data_2013_2018(folderpath,2016)
}
