% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_npf_event_clas_DM.R
\name{import_npf_event_clas_DM}
\alias{import_npf_event_clas_DM}
\title{import npf event clas DM}
\usage{
import_npf_event_clas_DM(where, filename)
}
\arguments{
\item{where}{path to the folder containing the file}

\item{filename}{name of the file with extension}
}
\value{
data table of the clasification file
}
\description{
Imports the npf-event clasification file created by Diego Aliaga and ALkuin Koenig
}
\examples{
import_npf_event_clas_DM(yourplace,"event_clasification_Diego_Max.csv")
}
