% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_SA_dat.R
\name{import_SA_dat}
\alias{import_SA_dat}
\title{Import SA data}
\usage{
import_SA_dat(where, filename)
}
\arguments{
\item{where}{path to file containing folder}

\item{filename}{filename with extension}
}
\value{
data in a data table
}
\description{
Imports sulfuric acid data derived by apitof for saltena campaign.
Data apparently in local
}
\examples{
import_SA_dat(filepath,"GEORGE_SA.csv" )

}
