% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_lm_eqn.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Add linear model equation}
\usage{
lm_eqn(df)
}
\arguments{
\item{df}{the dataframe. You calculate the model y~x}
}
\value{
an expression of the results of the linear model
}
\description{
This function adds the results of a linear model to a plot. It is deprecated by myself,
but might linger around in some code. I will remove it at some point.
}
\examples{
none
}
