% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_normalizeIt.R
\name{normalizeIt}
\alias{normalizeIt}
\title{normalize it}
\usage{
normalizeIt(vec, how)
}
\arguments{
\item{vec}{the (numerical) vector to be normalized}

\item{how}{how to normalize. Accepted: "norma","max","median","mean"}
}
\value{
the normalized vector
}
\description{
a function offering different ways of normalizing a vector
}
\examples{
normalizeIt(c(5,4,10,11,4),"norma")
normalizeIt(c(5,4,10,11,4),"max")
}
