% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_grid_extended.R
\name{plot_grid_extended}
\alias{plot_grid_extended}
\title{plot_grid_extended}
\usage{
plot_grid_extended(plotlist, relative_heights = rep(1, length(plotlist)),
  relative_widths = c(1, 0.5))
}
\arguments{
\item{plotlist}{list of plots to arrange. By default, align="hv"}

\item{relative_heights}{relative heights of the plots (rel_heights argument in plot_grid).
The length of this vector is equal to the number of plots in the grid}

\item{relative_widths}{relative widths of plots vs legends. The length of this vector
is always 2.}
}
\value{
an arranged plot
}
\description{
Allows arranging a list of plots with differing/partially non existing legends into
a single plot.
}
\details{
Extends upon the plot_grid function in the cowplot package. It will
create one plot grid for the plots and one for the legends and then stick those together
}
\examples{
plot_grid_extended(list(p1,p2,p3),c(1,1,0.5),c(1,0.3))
}
