% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotGR_splined_compa.R
\name{plotGR_splined_compa}
\alias{plotGR_splined_compa}
\title{Plot GR splined}
\usage{
plotGR_splined_compa(splined_rep_df, npfevent_size_frame)
}
\arguments{
\item{splined_rep_df}{Dataframe of evaluated splined functions}

\item{npfevent_size_frame}{data frame containing the growing modes}
}
\value{
a ggplot object
}
\description{
plots GR obtained by the splined approach. Binning is specific (can be found within the fun)
}
\details{
npfevent_size_frame only used as mask to check which trajectories to plot.
}
\examples{
plotGR_splined_compa(splined_rep_df,npfevent_size_frame)
}
