% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_GR_classic_box.R
\name{plotGRclassic}
\alias{plotGRclassic}
\title{Plot GR classic}
\usage{
plotGRclassic(npfevent_size_frame, event_GR_long, eventIDclas)
}
\arguments{
\item{npfevent_size_frame}{GR information}

\item{event_GR_long}{GR calculated}

\item{eventIDclas}{clasification of events}
}
\value{
A ggplot object
}
\description{
Plot growth rates obtained with the classic approach
}
\details{
This description has to be updated
}
\examples{
None
}
