% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_NAIS_nds.R
\name{plotNAIS_nds}
\alias{plotNAIS_nds}
\title{plot NAIS nds}
\usage{
plotNAIS_nds(data)
}
\arguments{
\item{data}{data to be plotted}
}
\value{
a ggplot object
}
\description{
plots nais data in nds format
}
\examples{
none
}
